Split failing ActionVIOnCompile test (#32)
The ActionVIOnCompile test added by #26 passed on the pull request, but has been failing on master. This re-writes the test to explicitly only call the code we want to validate, and avoids calling code that requires VeriStand to be running. It also moves the test into a new test case, to make testing future compile subVIs easier.