Error when handling a nonintegral state (#59)
Previously we were converting a double to an int32, but not checking that the conversion was lossless. We now confirm that the originally request state is an integer before proceeding.

Add tests to confirm the new behavior works as expected, and move Create Messages tests into a new test case.