Remove persisted resource from SLSC Switch custom device (#121)
Removes the resource property from the SLSC Switch custom device. We use the SLSC resource provided to us by VeriStand on deployment.

The hardware API has been modified to return the hardware custom device path as the resource. This is temporary until we can remove that accessor fully. It should be sufficient for now to prevent breakages, as the path is unique per custom device and deterministic.

We want to move away from using the SLSC resource in Routing and Faulting. There are other changes in flight for the Routing and Faulting custom device, so the scope of this change is currently limited to the SLSC Switch custom device as much as possible.