Update import/export tests
The tests are currently failing, but this is due to issues with the underlying (de)serializers used, specifically the Advanced System Definition API.

With these bugs fixed in my local install, most of the tests pass, with the remaining issue being that we are not correctly clearing out existing data when importing. To implement this correctly, changes are needed to the import/export tool.