Single source module paths, allowing for better rename handling (#152)
Change the persisted format for aliases and connections to no longer refer to a module directly. Instead, the aliases and connections refer to a unique module identifier (currently a number), which allows us to use dependent node properties to automatically track path changes.

Update the aliases scripting accessors to use a cluster, instead of a concatenated string for the endpoint name. This makes them more consistent with the state configuration workflow, and simplifies the implementation of the above change.

We currently do not do any cleanup of the module properties when they are no longer referenced; this can be changed in a backwards compatible way in the future.

The import/export tests are currently failing, but this is due to issues with the underlying (de)serializers used, specifically the Advanced System Definition API.

With these bugs fixed in my local install, most of the tests pass, with the remaining issue being that we are not correctly clearing out existing data when importing. To implement this correctly, changes are needed to the import/export tool.