Update the palettes and scripting APIs to address API review feedback (#172)
This addresses a large portion of the feedback from the API review process. This includes significant changes to the palette structure, as well as a number of VI renames.

Not all feedback is addressed in this review, in order to keep the number of changes remotely manageable. 

The following feedback has not yet been addressed and is being deferred to a different pull request, as the majority of the VIs in the repository will be affected:
- Remove the `ref` suffix from parameter names
- Remove the `(no error)` suffix from error in parameter names
- Use lower case (or at least consistent case) parameter names
- Use the VeriStand item ref typedef everywhere, instead of raw U64s

Additionally, the examples will need to be reviewed and updated in a separate pull request.