Support setting reservation group property for SLSC Switch (#196)
Adds support for setting the reservation group property added by NI-SLSC Switch 20.0. This allows for multiple concurrent reservations with the NI-SLSC driver.

The property is added to the system definition, and by default (empty string property) the path to the custom device instance is used for the reservation group. If the user.CD.Reservation Group property is set to a non-empty value, that value will be used as the reservation group instead.

There is currently not support for overriding this property to a different value in the UI, as this is expected to be a niche use case. However, the configured reservation group is displayed.

There should be no functional difference between having and not-having a reservation group in the case of a single accessor. Since NI-SLSC Switch does not hold a reservation throughout the entire session, and reserves/unreserves dynamically, customers could not have been relying on the lack of a reservation group as a security mechanism, so this change is safe to make.