Create NI-SWITCH custom device (#220)
Create an NI-SWITCH implementation of the Routing and Faulting consumer, based on the existing SLSC Switch implementation.

This change contains the minimal modifications to get NI-SWITCH working with R+F, while maintaining some semblance of coherency in the implementation. It won't work with NI-SWITCH modules that have NC relays, nor does it contain any automated tests or significant UI changes.

I tried to stay as close as possible to the existing SLSC Switch implementation to make reviewing easier, which means there are some rough edges. Minutiae such as error codes and context help will be updated in a subsequent commit.