Clean up R+F scripting error codes (#237)
Although changing error codes is a breaking change, the error codes currently used by Routing and Faulting are painfully inconsistent, to the point that it is nearly impossible to be successful handling them in an application. This one time breakage should make it significantly easier for customers to handle errors in the future.